+++ view/frontend/templates/search/page.phtml
@@ -2,18 +2,18 @@
 
 use Magento\Framework\Escaper;
 
-/** @var \Tweakwise\Magento2Tweakwise\Block\Catalog\PersonalMerchandisingAnalytics $block */
-
-$searchQuery = $block->getSearchQuery();
+/** @var \Tweakwise\Magento2Tweakwise\ViewModel\PersonalMerchandisingAnalytics $viewModel */
+$viewModel = $block->getData('viewModel');
 ?>
 
 <script type="text/x-magento-init">
     {
         "*": {
-            "Tweakwise_Magento2Tweakwise/js/search-page": {
-              "instanceKey": "<?= $escaper->escapeHtmlAttr($instanceKey); ?>"
+            "Tweakwise_Magento2Tweakwise/js/analytics": {
+              "searchQuery": "<?= $escaper->escapeHtmlAttr($viewModel->getSearchQuery()); ?>",
+              "type": "search"
         }
     }
 }
 </script>
-<script src="<?= $escaper->escapeHtmlAttr($block->getViewFileUrl('Tweakwise_Magento2Tweakwise::js/search-page.js')); ?>"></script>
+<script src="<?= $escaper->escapeHtmlAttr($block->getViewFileUrl('Tweakwise_Magento2Tweakwise::js/analytics.js')); ?>"></script>
